split SPI module into cloud and data-transfer
