+++ extensions/cloud/portability-cloud-google/build.gradle
@@ -1,4 +1,5 @@
 
 dependencies {
-    compile project(':portability-spi')
+    compile project(':portability-spi-data-transfer')
+    compile project(':portability-spi-cloud')
 }
