+++ portability-spi-cloud/build.gradle
