+++ portability-spi-cloud/src/main/java/org/dataportabilityproject/spi/client/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Defines the data portability client SPI. Extensions use this module to implement custom functionality for the system.
+ */
+package org.dataportabilityproject.spi;
