+++ portability-spi-data-transfer/build.gradle
@@ -1,0 +1,5 @@
+
+dependencies {
+    compile project(':portability-model')
+
+}
