+++ portability-spi-data-transfer/src/main/java/org/dataportabilityproject/spi/datatransfer/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Defines the data portability SPI. Extensions use this module to implement custom functionality for the system.
  */
-package org.dataportabilityproject.spi;
+package org.dataportabilityproject.spi.datatransfer;
